Add a logging adapter to append a correlation id
Add a class called CorrelationLoggerAdapter which extends
`logging.LoggerAdapter`.  When provided with extra information
containing an a `correlation_id` key, the adapter will append the id to
each log message sent with the adapter.

Removes py2.7 from tox's environment and from the Travis CI definition.

Runs the formatter black for the python code in this project.